requirements.txt
adrf==0.1.9
amqp==5.3.1
asgiref==3.9.1
async-property==0.2.2
attrs==25.3.0
billiard==4.2.1
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.5
django-cors-headers==4.7.0
django-redis==6.0.0
djangorestframework==3.16.1
drf-spectacular==0.28.0
h11==0.16.0
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
redis==6.4.0
referencing==0.36.2
rpds-py==0.27.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
uvicorn==0.35.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
